ingest/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prorag-ingest"
version = "0.1.0"
description = "pro-rag ingestion pipeline"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "sentence-transformers>=3.0",
    "tiktoken>=0.7",
    "click>=8.1",
    "python-docx>=1.1",
    "beautifulsoup4>=4.12",
    "pdfplumber>=0.11",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
prorag-ingest = "ingest.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py311"
line-length = 120
